Agregar Requerimiento 3

+++ src/main/java/com/api/demo/services/PersonaService.java
@@ -91,7 +91,7 @@ public class PersonaService {
         String password = generatePassword();
         String apikey = generateApiKey();
 
-        UsuarioId usuarioId = new UsuarioId(persona.getId(), login);
+        UsuarioId usuarioId = new UsuarioId((long)persona.getId(), login);
         usuario.setId(usuarioId);
         usuario.setPersona(persona);
         usuario.setPassword(password);
